[BUGFIX] Execute FreeSWITCH output components within the call actor
Avoids creating further actors for relatively inactive output components. Will reduce thread count and hopefully maximum concurrent call volume. If this works, we'll do the same thing for Input/Record and perhaps Asterisk.
