+++ spec/punchblock/translator/freeswitch/component/tts_output_spec.rb
@@ -37,7 +37,7 @@ module Punchblock
           describe '#execute' do
             before { original_command.request! }
             def expect_playback(voice = default_voice)
-              subject.wrapped_object.should_receive(:application).once.with :speak, "#{media_engine}|#{voice}|#{ssml_doc}"
+              subject.should_receive(:application).once.with :speak, "#{media_engine}|#{voice}|#{ssml_doc}"
             end
 
             let :ssml_doc do
@@ -255,13 +255,13 @@ module Punchblock
               end
 
               it "sets the command response to true" do
-                subject.wrapped_object.should_receive(:application)
+                subject.should_receive(:application)
                 subject.execute_command command
                 command.response(0.1).should be == true
               end
 
               it "sends the correct complete event" do
-                subject.wrapped_object.should_receive(:application)
+                subject.should_receive(:application)
                 original_command.should_not be_complete
                 subject.execute_command command
                 reason.should be_a Punchblock::Event::Complete::Stop
@@ -269,7 +269,7 @@ module Punchblock
               end
 
               it "breaks the current dialplan application" do
-                subject.wrapped_object.should_receive(:application).once.with 'break'
+                subject.should_receive(:application).once.with 'break'
                 subject.execute_command command
               end
             end
